refactor(lb): added tests for backend pool
